Work on summation inference (#49)
* Clean up summation infer

This will now only propagate form types. If a single form is present then this will be inferred. If there are no forms then we skip and if there are multiple forms we error.

* Loop only through vars to infer

* Fixed error introduced

* Moved infer sums loop

This brings its function signature more in line with the other inference applications.

* Better error for sum type mismatch

* Turn on op2 lookup

* Touching up

* A final test

* Clarifying filterfor_forms

* Better type group handling

I'm testing the use of getter functions to return us type groups to avoid us having to type them out. We can add more groups if needed.

* Changed macros to const